commit 17 changes form structure and logic to use primary key instead of student
Added new id field (Long) in Student model for JPA operations before was using studentid as for jpa operation getting problems change logic
Updated controller and service logic to use id for editing and deleting
 Modified html templates form.html and list.html to work with 'id' instead of studentId
Added backend validation for studentId and graduationYear

Styled and added modal popup for unsaved form confirmation Back button
Created student_form.js to handle custom modal logic
Improved style by adding placeholders to all fields
 Created student_form.js to handle custom Back to List confirmation using modal replaces browser popup
Connected student_form.js in form.html for modular JavaScript support
All styles and modal layout handled through existing CSS

+++ student_Crud/student_Crud/src/main/java/com/example/student_Crud/controller/StudentController.java
@@ -21,6 +21,7 @@ import jakarta.validation.Valid;
 public class StudentController {
 	@Autowired
     private StudentService studentService;
+	
 
 	@GetMapping("/list")
 	public String studentlist(Model model) {
@@ -53,14 +54,14 @@ public class StudentController {
 
     
     @GetMapping("/list/delete/{id}")
-    public String deleteStudent(@PathVariable("id") int id) {
+    public String deleteStudent(@PathVariable("id") long id) {
     	  studentService.deleteStudentById(id);
         return "redirect:/list";
     }
     
     
     @GetMapping("/list/edit/{id}")
-    public String editStudent(@PathVariable("id") int id, Model model) {
+    public String editStudent(@PathVariable("id") long id, Model model) {
         Student student = studentService.getStudentById(id);
         model.addAttribute("student", student);
         return "student_form";
@@ -69,6 +70,7 @@ public class StudentController {
     public String loginPage() {
         return "login"; 
     }
+  
 
-
+  
 }
