+++ student_Crud/student_Crud/src/main/java/com/example/student_Crud/service/StudentService.java
@@ -9,6 +9,6 @@ public interface StudentService {
 	
     List<Student> getAllStudents();
     void saveStudent(Student student);
-    Student getStudentById(int id);
-    void deleteStudentById(int id);
+    Student getStudentById( long id);
+    void deleteStudentById(long id);
 }
